BUG-41848: IoT - Storage Layer - JDBC implementation     1. Refactor code to cache prepared statements (NOT Compiling)

+++ core/src/main/java/com/hortonworks/iotas/storage/impl/jdbc/mysql/query/MySqlDelete.java
@@ -27,11 +27,16 @@ public class MySqlDelete extends MySqlStorableKeyBuilder {
 
     // "DELETE FROM DB.TABLE WHERE id1 = val1 AND id2 = val2"
     @Override
-    public String getParameterizedSql() {
-
-        final String sql = "DELETE FROM  " + tableName + " WHERE "
-                + join(getColumnNames(columns, "%s = ?"), " AND ");
-        log.debug(sql);
-        return sql;
+    public String getParametrizedSql() {
+        if (paramSql == null) {
+            final String sql;
+            sql = "DELETE FROM  " + tableName + " WHERE "
+                    + join(getColumnNames(columns, "%s = ?"), " AND ");
+            paramSql = sql;
+            log.debug(sql);
+        }
+        return paramSql;
     }
+
+
 }
